sdks/akta-docs/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "akta-docs"
version = "0.1.0"
description = "OPENAKTA documentation linter and scaffolding (GEO / AI context)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "OPENAKTA" }]
keywords = ["openakta", "markdown", "linter", "documentation", "llm"]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "python-frontmatter>=1.1.0",
  "markdown-it-py>=3.0.0",
  "pathspec>=0.12.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8.0", "build>=1.0"]

[project.scripts]
akta-docs = "akta_docs.cli:main"

[tool.hatch.build.targets.sdist]
include = ["src/akta_docs", "README.md", "pyproject.toml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
